+++ receiver/kafkareceiver/unmarshaler.go
@@ -16,6 +16,9 @@ package kafkareceiver
 
 import (
 	"go.opentelemetry.io/collector/consumer/pdata"
+	"go.opentelemetry.io/collector/internal/otlp"
+	"go.opentelemetry.io/collector/translator/trace/zipkinv1"
+	"go.opentelemetry.io/collector/translator/trace/zipkinv2"
 )
 
 // TracesUnmarshaler deserializes the message body.
@@ -24,6 +27,15 @@ type TracesUnmarshaler interface {
 	Unmarshal([]byte) (pdata.Traces, error)
 
 	// Encoding of the serialized messages.
+	Encoding() string
+}
+
+// MetricsUnmarshaler deserializes the message body
+type MetricsUnmarshaler interface {
+	// Unmarshal deserializes the message body into traces
+	Unmarshal([]byte) (pdata.Metrics, error)
+
+	// Encoding of the serialized messages
 	Encoding() string
 }
 
@@ -38,14 +50,14 @@ type LogsUnmarshaler interface {
 
 // defaultTracesUnmarshalers returns map of supported encodings with TracesUnmarshaler.
 func defaultTracesUnmarshalers() map[string]TracesUnmarshaler {
-	otlp := &otlpTracesPbUnmarshaler{}
+	otlpPb := newPdataTracesUnmarshaler(otlp.NewProtobufTracesUnmarshaler(), defaultEncoding)
 	jaegerProto := jaegerProtoSpanUnmarshaler{}
 	jaegerJSON := jaegerJSONSpanUnmarshaler{}
-	zipkinProto := zipkinProtoSpanUnmarshaler{}
-	zipkinJSON := zipkinJSONSpanUnmarshaler{}
-	zipkinThrift := zipkinThriftSpanUnmarshaler{}
+	zipkinProto := newPdataTracesUnmarshaler(zipkinv2.NewProtobufTracesUnmarshaler(false, false), "zipkin_proto")
+	zipkinJSON := newPdataTracesUnmarshaler(zipkinv2.NewJSONTracesUnmarshaler(false), "zipkin_json")
+	zipkinThrift := newPdataTracesUnmarshaler(zipkinv1.NewThriftTracesUnmarshaler(), "zipkin_thrift")
 	return map[string]TracesUnmarshaler{
-		otlp.Encoding():         otlp,
+		otlpPb.Encoding():       otlpPb,
 		jaegerProto.Encoding():  jaegerProto,
 		jaegerJSON.Encoding():   jaegerJSON,
 		zipkinProto.Encoding():  zipkinProto,
@@ -54,9 +66,16 @@ func defaultTracesUnmarshalers() map[string]TracesUnmarshaler {
 	}
 }
 
+func defaultMetricsUnmarshalers() map[string]MetricsUnmarshaler {
+	otlpPb := newPdataMetricsUnmarshaler(otlp.NewProtobufMetricsUnmarshaler(), defaultEncoding)
+	return map[string]MetricsUnmarshaler{
+		otlpPb.Encoding(): otlpPb,
+	}
+}
+
 func defaultLogsUnmarshalers() map[string]LogsUnmarshaler {
-	otlp := &otlpLogsPbUnmarshaler{}
+	otlpPb := newPdataLogsUnmarshaler(otlp.NewProtobufLogsUnmarshaler(), defaultEncoding)
 	return map[string]LogsUnmarshaler{
-		otlp.Encoding(): otlp,
+		otlpPb.Encoding(): otlpPb,
 	}
 }