+++ receiver/kafkareceiver/unmarshaler_test.go
@@ -41,6 +41,21 @@ func TestDefaultTracesUnMarshaler(t *testing.T) {
 	}
 }
 
+func TestDefaultMetricsUnMarshaler(t *testing.T) {
+	expectedEncodings := []string{
+		"otlp_proto",
+	}
+	marshalers := defaultMetricsUnmarshalers()
+	assert.Equal(t, len(expectedEncodings), len(marshalers))
+	for _, e := range expectedEncodings {
+		t.Run(e, func(t *testing.T) {
+			m, ok := marshalers[e]
+			require.True(t, ok)
+			assert.NotNil(t, m)
+		})
+	}
+}
+
 func TestDefaultLogsUnMarshaler(t *testing.T) {
 	expectedEncodings := []string{
 		"otlp_proto",